Show attached files in the search results
