+++ components/PostsList.js
@@ -2,17 +2,14 @@ import React from "react";
 
 import PostSummary from "./PostSummary";
 
-export default function PostsList({ posts, limit }) {
+export default function PostsList({ posts, limit, showAttachments }) {
   return posts
     .slice(0, Math.min(posts.length, limit || posts.length))
     .map((post) => (
       <PostSummary
         key={post.id}
-        id={post.id}
-        title={post.title}
-        summary={post.summary}
-        files={post.files}
-        date={new Date(post.created_at)}
+        post={post}
+        showAttachments={showAttachments}
       />
     ));
 }
