+++ screens/AppNavigation.js
@@ -5,8 +5,6 @@ import { createStackNavigator } from "@react-navigation/stack";
 
 import Home from "./Home.js";
 import Search from "./Search.js";
-import SearchPosts from "./SearchPosts";
-import SearchFiles from "./SearchFiles";
 import Question from "./Question.js";
 import QuestionSubmitted from "./QuestionSubmitted.js";
 import PostUpdate from "./PostUpdate.js";
@@ -49,8 +47,6 @@ export default function AppNavigation() {
     >
       <StackNavigator.Screen name="Home" component={Home} />
       <StackNavigator.Screen name="Search" component={Search} />
-      <StackNavigator.Screen name="SearchPosts" component={SearchPosts} />
-      <StackNavigator.Screen name="SearchFiles" component={SearchFiles} />
       <StackNavigator.Screen name="Updates" component={Updates} />
       <StackNavigator.Screen name="Question" component={Question} />
       <StackNavigator.Screen
