+++ screens/Search.js
@@ -1,27 +1,23 @@
 import React, { useRef, useEffect, useState, useCallback } from "react";
-import {
-  View,
-  TouchableWithoutFeedback,
-  Keyboard,
-  StyleSheet,
-} from "react-native";
-import { Appbar, Searchbar, ProgressBar } from "react-native-paper";
+import { View, TouchableWithoutFeedback, Keyboard } from "react-native";
+import { Appbar, Searchbar, ProgressBar, Text } from "react-native-paper";
 
-import PostsListWithButton from "../components/PostsListWithButton";
-import SectionWithButton from "../components/SectionWithButton";
-import AttachmentsList from "../components/AttachmentsList";
+import PostsList from "../components/PostsList.js";
+import InfiniteScrollView from "../components/InfiniteScrollView.js";
 
 import api from "../util/api";
 
-export default function Search({ navigation }) {
-  const DEFAULT_RESULTS = 4;
+export default function SearchPosts({ navigation }) {
+  const DEFAULT_SEARCH_LIMIT = 10;
+  const fullHeight = { flex: 1 };
 
   const ref = useRef(null);
   const [firstFocus, setFirstFocus] = useState(true);
-  const [searchText, setSearchText] = useState(true);
+  const [searchText, setSearchText] = useState("");
   const [foundPosts, setFoundPosts] = useState([]);
-  const [foundFiles, setFoundFiles] = useState([]);
+  const [limit, setLimit] = useState(DEFAULT_SEARCH_LIMIT);
   const [loading, setLoading] = useState(false);
+  const [loadedAll, setLoadedAll] = useState(false);
 
   useEffect(() => {
     return navigation.addListener("focus", () => {
@@ -32,42 +28,67 @@ export default function Search({ navigation }) {
     });
   }, [firstFocus]);
 
+  useEffect(() => {
+    update(searchText);
+  }, []);
+
   const update = useCallback(
-    ({ text }) => {
-      if (text === "") {
+    ({ text, loadMore }) => {
+      if (text === "" || (text === undefined && searchText == "")) {
         // Return [] when searching for an empty strings
         setFoundPosts([]);
-        setFoundFiles([]);
         return;
       }
 
       let ignoreOutdated = false;
+      // Keep limit in a temporary variable so that it can be updated synchronously
+      let currentLimit = limit;
+
+      if (loadMore === true && !loading) {
+        if (foundPosts.length < currentLimit) {
+          // No more posts to load
+          setLoadedAll(true);
+          return;
+        }
+        currentLimit = currentLimit + DEFAULT_SEARCH_LIMIT;
+        setLimit(currentLimit);
+      } else if (!loadMore) {
+        /* Search triggered through change of search term,
+           reset limit and carry on with search */
+        currentLimit = DEFAULT_SEARCH_LIMIT;
+        setLimit(currentLimit);
+        setLoadedAll(false);
+      } else if (loading) {
+        /* Search triggered through reaching end of the list,
+           but other request is pending - cancel */
+        return;
+      }
       setLoading(true);
 
-      async function search(text) {
-        const postsResults = await api.searchPosts(text, 0, DEFAULT_RESULTS);
-        const fileResults = await api.searchFiles(text, 0, DEFAULT_RESULTS);
-        if (!postsResults.success || !fileResults.success) {
+      async function search(text, fetchNumber) {
+        const results = await api.searchPosts(text, 0, fetchNumber);
+        if (!results.success) {
           return;
         }
         if (!ignoreOutdated) {
           // Update of found posts was not cancelled by more recent search
-          setFoundPosts(postsResults.data);
-          setFoundFiles(fileResults.data);
+          setFoundPosts(results.data);
         }
       }
 
-      search(text).then(() => setLoading(false));
+      search(text ? text : searchText, currentLimit).then(() =>
+        setLoading(false)
+      );
       return () => {
         // Cancel update of found posts
         ignoreOutdated = true;
       };
     },
-    [foundPosts, loading]
+    [foundPosts, searchText, limit, loading]
   );
 
   return (
-    <View style={styles.fullHeight}>
+    <View style={fullHeight}>
       <TouchableWithoutFeedback onPress={Keyboard.dismiss} accessible={false}>
         <Appbar.Header>
           <Appbar.BackAction onPress={() => navigation.goBack()} />
@@ -75,6 +96,7 @@ export default function Search({ navigation }) {
             <Searchbar
               placeholder="Search"
               style={{ height: 42 }}
+              value={searchText}
               ref={ref}
               onChangeText={(text) => {
                 setSearchText(text);
@@ -84,41 +106,20 @@ export default function Search({ navigation }) {
           </View>
         </Appbar.Header>
       </TouchableWithoutFeedback>
-      <View style={[styles.fullHeight, styles.contentContainer]}>
+      <View style={fullHeight}>
         <TouchableWithoutFeedback onPress={Keyboard.dismiss} accessible={false}>
-          <View>
-            <PostsListWithButton
-              title="Found posts"
-              buttonText="More"
-              onButtonPress={() =>
-                navigation.navigate("SearchPosts", { searchText: searchText })
-              }
-              loading={loading}
-              limit={4}
-              posts={foundPosts}
-            />
-            <SectionWithButton
-              title="Related files"
-              buttonText="More"
-              onButtonPress={() =>
-                navigation.navigate("SearchFiles", { searchText: searchText })
-              }
-            >
-              <AttachmentsList filesWithPosts={foundFiles} />
-            </SectionWithButton>
+          <InfiniteScrollView
+            onEndReached={() => update({ loadMore: true })}
+            onEndReachedThreshold={30}
+          >
+            <PostsList posts={foundPosts} showAttachments={true} />
+            {loadedAll && (
+              <Text style={{ textAlign: "center" }}>No more results</Text>
+            )}
             {loading && <ProgressBar indeterminate={true} />}
-          </View>
+          </InfiniteScrollView>
         </TouchableWithoutFeedback>
       </View>
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  contentContainer: {
-    padding: 16,
-  },
-  fullHeight: {
-    flex: 1,
-  },
-});
